fix dtype (#67)

+++ src/compressed_tensors/quantization/lifecycle/apply.py
@@ -203,4 +203,4 @@ def _load_quant_args_from_state_dict(
         if zp_from_state is not None:  # load the non-zero zero points
             zp.data = state_dict[f"{module_name}.{zp_name}"].to(device)
         else:  # fill with zeros matching scale shape
-            zp.data = torch.zeros_like(scale, dtype=int).to(device)
+            zp.data = torch.zeros_like(scale, dtype=torch.int8).to(device)
